+++ src/general.hpp
@@ -6,7 +6,7 @@
 /*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/09 17:35:25 by jvacaris          #+#    #+#             */
-/*   Updated: 2023/04/16 19:45:50 by jvacaris         ###   ########.fr       */
+/*   Updated: 2023/04/20 20:42:37 by jvacaris         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,6 +28,7 @@
 #include <ctime>
 #include <sys/stat.h>
 #include <sys/types.h>
+#include <dirent.h>
 //*		~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 //*		~~~~~~~ Text coloring ~~~~~~~
@@ -62,6 +63,6 @@ struct s_request_info
 };
 
 std::vector <std::string> cpp_split(std::string full_line, char splitter);
-std::string file_reader(const std::string &filename);
+std::string file_reader(const std::string &filename, int *status);
 
 #endif