+++ src/utils.cpp
@@ -6,7 +6,7 @@
 /*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/06 20:30:58 by jvacaris          #+#    #+#             */
-/*   Updated: 2023/04/16 19:46:51 by jvacaris         ###   ########.fr       */
+/*   Updated: 2023/04/19 18:34:54 by jvacaris         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,7 +41,7 @@ std::vector <std::string> cpp_split(std::string full_line, char splitter)
 	return (line_vector);
 }
 
-std::string file_reader(const std::string &filename)
+std::string file_reader(const std::string &filename, int *status)
 {
 	std::string		body;
 	std::ifstream	file_instream;
@@ -50,17 +50,21 @@ std::string file_reader(const std::string &filename)
 
 	if (!file_instream.is_open())
 	{
-		//	COMPLAIN
+		if (file_instream.fail())
+		{
+			*status = errno;
+		}
 		return ("");
 	}
 	else
 	{
+		*status = 0;
 		while(std::getline(file_instream, file_content_buff))
 		{
 			body.append(file_content_buff);
 			body.append("\n");
 		}
+		file_instream.close();
 		return (body);
 	}
-	file_instream.close();
 }